add QueryRegistrationHandler
